backend/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv